requirements.txt
google-generativeai
asyncio
dataclasses
enum34
typing
numpy
pandas
scikit-learn
matplotlib
seaborn
textblob
nltk
spacy
transformers
torch
websocket-client
flask
flask-socketio
openai
python-dotenv